playable

+++ app/src/main/java/com/example/hippoplayer/utils/ConvertHelper.java
@@ -3,7 +3,7 @@ package com.example.hippoplayer.utils;
 import java.util.concurrent.TimeUnit;
 
 public class ConvertHelper {
-    public static String convertToMinutes(int miliseconds) {
+    public static String convertToMinutes(long miliseconds) {
         long minutes = TimeUnit.MILLISECONDS.toMinutes(miliseconds);
         long seconds = TimeUnit.MILLISECONDS.toSeconds(miliseconds) - minutes * 60;
         String stringSecond = "";
